concealable frontlayer when that's active by pressing back button on phone
By default, when the frontLayer widget  visibility is active, the user expects to hide it by pressing the back button. However, in the current version, this did not happen, and unfortunately, outside of this library, I was unable to hide the frontLayer widget. To achieve this, I removed the ‍concealBacklayerOnBackButton‍ argument and modified the Scaffold class in this library. Now, this operation is successfully executed.